Update exercice 20 function name
Update exercice 20 function name to be aligned with the description (scaladoc above it).

The description (and also the solution) specifies that the effect should be executed in recursion UNTIL the condition returns true. So it is a doUntil instead of doWhile. In the later, the description could be WHILE the condition returns true, the effect should recurse.

Otherwise, the description should be updated if the function name can't be changed.